bring the frame Z back to 55mm, for taller levers
50mm - 10mm of panels was fine for some levers but not enough for the
LSX-NOBI, which I personally have a number of. I can't vet every lever
in existence, but I know at least that this was a problem that is solved
by adding 5mm, which seems like a fair trade and not a huge difference
in the final product.

this also helps the fact that at 40mm internal, some buttons + wiring
harnesses were getting pretty tight, mostly relevant at the frame
connection point where the inner bottom panel isn't a basin. this should
help that too.

additionally this has us go back to the 45mm M4 standoff recommendation,
which kinda sucks because it seems like 40mm are easier to find, but as
stated in the notes, the difference is somewhat negligible if you just
use 40mm with longer bolts.