frames: re-bevel the bottom and top, center buttons
this undoes my work to simplify the bottom of the frame and move button
holes in the frames accordingly. the top and bottom are now symmetrical
in terms of their bevel, and the buttons are centered. this will make
the print of the frame difficult again, but I have an upcoming change to
modularize the frame walls and that will print far better, so I'm
thinking this is what I want. testing ongoing

This reverts commit ed031b9308b062b8900a40851dafac5f18aaa535.
This reverts commit f90fc095e04b46073f33681b6cbe94b3ac421256.

Signed-off-by: Brian S. Stephan <bss@incorporeal.org>