favor symmetry in the front/back and side wall windowed pieces
I think the lack of symmetry led to a varying amount of force being
applied when joined with the interconnect. take for instance:

OUTER  INTER  INNER
INTER  INTER  INTER
INNER  INTER  OUTER

if OUTER (the "windowed" piece, the outside wall + one half of the
mounting column) is on top on the left side and on the bottom on the
right side, and INNER (the inner wall, the one that floats around
relative to the outer wall) vice versa on the other side, the two pieces
side by side would not create a clean line --- left would be ~0.5mm
lower than the right.

if there was a way to clamp these down relative to each other, the
problem might go away, but these are pretty pivotal to the overall stick
being aligned and the panels don't cross the interconnect, so probably
panels and a completed stick would have been similarly misaligned.

this, I believe, fixes it, and also fixes a tiny issue where a bit too
much material was being cut out of the inner walls by way of the mount
columns having a triangular point not normally seen because it's inside
the inner wall

Signed-off-by: Brian S. Stephan <bss@incorporeal.org>