retain a bit of the inner wall on the outer piece for windowed pieces
the inner wall piece being the *whole* inner wall of the frame side
meant that the (now relatively thin) outer wall piece was not supported
by anything on the top and bottom, which combined with the already weak
piece because of the window, gave the edges a lot of give.

shrinking the inner wall piece slightly retains some of the material on
the outer piece, which should friction fit the top and bottom with both
the inner wall piece, and with the panels, leading to a lot more clamp
when everything is secured.

Signed-off-by: Brian S. Stephan <bss@incorporeal.org>