create a panel with three aux buttons in a tight cluster
the idea of this is to have Select, Home, and Start buttons accessible
via this panel away from the action panels, and the action pannels have
the L3, TP, and R3 buttons in the Sega 2P 11 layout, and then there is
no need for buttons on the back of the q2009 enclosure. I don't think
this would get used for any reason other than q2009 so it lives there.

Signed-off-by: Brian S. Stephan <bss@incorporeal.org>