ensure the frame solid has the proper, complete dimensions
something about the sharp corners of the default cylinder was slightly
shrinking the overall solid by a fraction of a mm. I thought this was a
defect of how I was chopping pieces, but no, it was the core solid. this
shouldn't be a reason to rebuild anything, but it makes the models
slightly more consistent, especially when looking at the STLs

Signed-off-by: Brian S. Stephan <bss@incorporeal.org>