use the derived frame piece y dimension to build side
as in previous commit, this simplifies the build of the side frame piece
a bit. notably on this one, the preview of the piece is more performant,
so I'm thinking this is a good move. the object triangle count and
volume are slightly different (not sure why), but the object dimensions
are the same

still didn't get rid of any 2-manifold errors though :(

Signed-off-by: Brian S. Stephan <bss@incorporeal.org>