add little braces to go on the corner and middle seams
I like the corners especially. these pieces slide over the seam points
on box sticks and give a bit of depth and can cover some sharp points in
the corners where the bevels meet. so far, I've printed these in the
same color as the frame, but you could mix and match to taste, of
course.

the piece that goes in the middle is not my favorite, but I designed it
and it works, so might as well commit it. maybe these would look better
with a pattern or logo etched in.

Signed-off-by: Brian S. Stephan <bss@incorporeal.org>