add a simple sleeve for a hex spacer that isn't wide enough
I got a bag of spacers that are not as wide as what I normally use, so
they rattle around in the post hole and obviously don't help with the
fit of the completed stick. this sleeve seems to be right for them to
fill the hole. provided in case it is useful in the future, for myself
or anyone else

Signed-off-by: Brian S. Stephan <bss@incorporeal.org>